@@ -1,75 +1,114 @@
 {
   "name": "lightbox-react",
-  "version": "0.1.6",
+  "version": "0.3.6",
   "description": "Lightbox for React components or images",
   "scripts": {
-    "build": "npm run lint && npm run build:demo && npm run build:umd",
-    "build:demo": "npm run clean:demo && cross-env NODE_ENV=production webpack --config webpack.config.demo.babel.js --bail",
-    "build:umd": "npm run clean:umd && cross-env NODE_ENV=production webpack --config webpack.config.umd.babel.js --bail",
-    "clean": "npm run clean:demo && npm run clean:umd",
+    "build": "npm run clean && cross-env NODE_ENV=production TARGET=umd webpack --bail",
+    "build:demo": "npm run clean:demo && cross-env NODE_ENV=production TARGET=demo webpack --bail",
+    "clean": "rimraf dist style.css style.css.map",
     "clean:demo": "rimraf build",
-    "clean:umd": "rimraf dist/umd",
-    "start": "cross-env NODE_ENV=development webpack-dev-server --hot --inline --config webpack.config.dev.babel.js",
-    "test": "NODE_ENV=production webpack --config webpack.config.umd.babel.js",
-    "lint": "eslint src",
-    "deploy": "npm run build && gh-pages -d build",
-    "prepublish": "npm run lint && npm run build:umd"
+    "start": "cross-env NODE_ENV=development TARGET=development webpack-dev-server --inline --hot",
+    "lint": "eslint .",
+    "prettier": "prettier --single-quote --trailing-comma es5 --write \"./**/*.{md,js,css}\" \"!./{build,dist}/**\" \"!./style.css*\"",
+    "prepublishOnly": "npm run lint && npm run test && npm run build",
+    "test": "jest",
+    "test:watch": "jest --watchAll",
+    "deploy": "npm run build:demo && gh-pages -d build"
   },
-  "main": "dist/umd/lightbox-react.js",
+  "main": "dist/main.js",
+  "typings": "index.d.ts",
   "files": [
-    "dist"
+    "dist",
+    "index.d.ts",
+    "style.css",
+    "style.css.map"
   ],
   "repository": {
     "type": "git",
     "url": "https://github.com/treyhuffine/lightbox-react.git"
   },
-  "homepage": "http://treyhuffine.com/lightbox-react/",
+  "homepage": "http://treyhuffine.com/lightbox-react",
   "bugs": "https://github.com/treyhuffine/lightbox-react/issues",
   "authors": [
     "Trey Huffine",
     "Chris Fritz <forked>"
   ],
   "license": "MIT",
+  "jest": {
+    "testURL": "http://localhost",
+    "setupTestFrameworkScriptFile": "./node_modules/jest-enzyme/lib/index.js",
+    "setupFiles": [
+      "./test-config/shim.js",
+      "./test-config/test-setup.js"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "jsx",
+      "json"
+    ],
+    "moduleDirectories": [
+      "node_modules"
+    ],
+    "moduleNameMapper": {
+      "\\.(css|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js"
+    },
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ]
+  },
+  "browserslist": [
+    "IE >= 11",
+    "last 2 versions",
+    "> 1%"
+  ],
   "dependencies": {
-    "is-react": "^1.0.0",
-    "react-modal": "^1.4.0"
+    "is-react": "^1.3.2",
+    "prop-types": "^15.6.1",
+    "react-modal": "^3.4.4"
   },
   "peerDependencies": {
-    "prop-types": "^15.5.8",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0"
+    "react": "^15.5.0 || ^16.0.0",
+    "react-dom": "^15.5.0 || ^16.0.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.3.7",
+    "autoprefixer": "^8.4.1",
     "babel-cli": "^6.10.1",
-    "babel-core": "^6.10.4",
-    "babel-eslint": "^7.1.0",
-    "babel-loader": "^6.2.4",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.4",
     "babel-plugin-transform-exponentiation-operator": "^6.8.0",
     "babel-plugin-transform-object-rest-spread": "^6.8.0",
-    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.11.1",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.26.0",
-    "eslint": "^3.2.2",
-    "eslint-config-blue-hour": "^0.2.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^6.0.0",
-    "file-loader": "^0.9.0",
-    "gh-pages": "^0.12.0",
-    "html-webpack-plugin": "^2.22.0",
-    "node-sass": "^3.8.0",
-    "postcss-loader": "^1.1.0",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-hot-loader": "^1.3.0",
+    "coveralls": "^3.0.1",
+    "cross-env": "^5.1.4",
+    "css-loader": "^0.28.11",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.3",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^1.1.11",
+    "gh-pages": "^1.1.0",
+    "html-webpack-plugin": "^3.2.0",
+    "jest": "^22.4.3",
+    "jest-enzyme": "^6.0.0",
+    "postcss-loader": "^2.1.5",
+    "prettier": "^1.12.1",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "react-hot-loader": "^4.1.2",
     "rimraf": "^2.5.3",
-    "sass-loader": "^4.0.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "style-loader": "^0.21.0",
+    "uglifyjs-webpack-plugin": "^1.2.5",
+    "webpack": "^4.11.0",
+    "webpack-cli": "^3.1.1",
+    "webpack-dev-server": "^3.1.4",
+    "webpack-node-externals": "^1.7.2"
   },
   "keywords": [
     "react",
